Move interfaces under consensus
