+++ packages/contracts/source/contracts/consensus.ts
@@ -19,6 +19,24 @@ export interface IConsensusService {
 	onTimeoutPrecommit(height: number, round: number): Promise<void>;
 }
 
+export interface IHandler {
+	onProposal(proposal: IProposal): Promise<void>;
+	onPrevote(prevote: IPrevote): Promise<void>;
+	onPrecommit(precommit: IPrecommit): Promise<void>;
+}
+
+export interface IBroadcaster {
+	broadcastProposal(proposal: IProposal): Promise<void>;
+	broadcastPrevote(prevote: IPrevote): Promise<void>;
+	broadcastPrecommit(precommit: IPrecommit): Promise<void>;
+}
+
+export interface IScheduler {
+	scheduleTimeoutPropose(height: number, round: number): Promise<void>;
+	scheduleTimeoutPrevote(height: number, round: number): Promise<void>;
+	scheduleTimeoutPrecommit(height: number, round: number): Promise<void>;
+}
+
 export interface IValidator {
 	configure(publicKey: string, keyPair: IKeyPair): IValidator;
 	getConsensusPublicKey(): string;
